pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Ирина Судейкина",email = "irinaasster@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"  # Новый диапазон Python

dependencies = [
    "django (>=4.0.1,<6.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "drf-yasg (>=1.21.14,<2.0.0)",
    "django-filter (>=25.2,<26.0)",
    "coverage (>=7.13.4,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "eventlet (>=0.39.4,<0.41.0)",
    "redis (>=6.1.1,<8.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "pytz (>=2025.2,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^26.1.0"
isort = "^7.0.0"



[tool.poetry.group.dev.dependencies]
coverage = "^7.13.4"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''


[tool.isort]
# максимальная длина строки
line_length = 119
